+++ ci/release_notes.md
@@ -1,1 +1,3 @@
-Empty - please add release notes here
+## Improvement
+
+- reproduce last state when there is nothing to prepare in ci_in